Bug #1517552 fix coverity bug
[root cause  ] uint32_t has a larger boundary than uint8_t
[changes     ] libcamera
[side effects] no
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: no
[is there dependence]: no
[confirm dependent commit]: yes
[board]: l6
[test case]: camera
[reviewers   ] victor.xu
[change_type ] feature_bugfix
[tag_product ] common

+++ hal3_2v6/SprdCamera3HWI.cpp
@@ -2234,7 +2234,7 @@ int SprdCamera3HWI::isStreamCombinationSupported(
         ret = -EINVAL;
         goto exit;
     }
-    for (uint8_t i = 0; i < comb->num_streams; i++) {
+    for (uint32_t i = 0; i < comb->num_streams; i++) {
         streamBuffer[i] = {comb->streams[i].stream_type,
                            comb->streams[i].width,
                            comb->streams[i].height,
